pyproject.toml
[project]
name = "pdf_summarizer_quiz_generator"
version = "0.1.0"
description = "AI-powered PDF Summarizer + Quiz Generator Agent"
dependencies = [
    "openai-agents",
    "pypdf",
    "python-dotenv",
    "reportlab>=4.4.5",
    "streamlit",
    "blinker",
    "protobuf",
    "typing_extensions",
]
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]
